rename proto lib
