+++ services/app-api/src/postgres/storeError.test.ts
@@ -2,7 +2,7 @@
 import { v4 as uuidv4 } from 'uuid'
 import { PrismaClient } from '@prisma/client'
 import { UnlockedHealthPlanFormDataType } from '../../../app-web/src/common-code/healthPlanFormDataType'
-import { toProtoBuffer } from '../../../app-web/src/common-code/proto/stateSubmission'
+import { toProtoBuffer } from '../../../app-web/src/common-code/proto/healthPlanFormDataProto'
 import { sharedTestPrismaClient } from '../testHelpers/storeHelpers'
 import { convertPrismaErrorToStoreError } from './storeError'
