+++ services/app-api/src/resolvers/submitHealthPlanPackage.test.ts
@@ -10,7 +10,7 @@ import {
     createAndSubmitTestHealthPlanPackage,
 } from '../testHelpers/gqlHelpers'
 import { testEmailConfig, testEmailer } from '../testHelpers/emailerHelpers'
-import { base64ToDomain } from '../../../app-web/src/common-code/proto/stateSubmission'
+import { base64ToDomain } from '../../../app-web/src/common-code/proto/healthPlanFormDataProto'
 import { packageName } from '../../../app-web/src/common-code/healthPlanFormDataType'
 import { latestFormData } from '../testHelpers/healthPlanPackageHelpers'
