+++ services/app-web/src/common-code/proto/healthPlanFormDataProto/index.ts
@@ -1,0 +1,3 @@
+export { base64ToDomain, domainToBase64, protoToBase64 } from './protoBase64'
+export { toDomain } from './toDomain'
+export { toProtoBuffer } from './toProtoBuffer'
