--- services/app-web/src/common-code/proto/stateSubmission/index.ts
@@ -1,7 +1,0 @@
-export {
-    base64ToDomain, domainToBase64,
-    protoToBase64
-} from './protoBase64'
-export { toDomain } from './toDomain'
-export { toProtoBuffer } from './toProtoBuffer'
-
